pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "augur"
version = "0.1.0"
description = "Multi-specialist ensemble forecasting — fan probabilistic questions to domain experts and aggregate calibrated predictions."
readme = "README.md"
license = { text = "FSL-1.1-Apache-2.0" }
requires-python = ">=3.11"
keywords = ["forecasting", "prediction", "ensemble", "llm", "superforecasting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.25",
    "openai>=1.30",
    "google-genai>=1.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["augur"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
